+++ rudof_cli/src/node.rs
@@ -1,14 +1,11 @@
 use anyhow::*;
 use iri_s::IriS;
 use rudof_lib::node_info::{NodeInfoOptions, get_node_info};
-use rudof_lib::{InputSpec, Rudof, RudofConfig};
+use rudof_lib::{InputSpec, Rudof, RudofConfig, format_node_info_list, parse_node_selector};
 use srdf::ReaderMode;
 use std::path::PathBuf;
 
-use crate::{
-    ShowNodeMode, data::get_data_rudof, data_format::DataFormat,
-    node_formatter::format_node_info_list, node_selector::parse_node_selector, writer::get_writer,
-};
+use crate::{ShowNodeMode, data::get_data_rudof, data_format::DataFormat, writer::get_writer};
 
 #[allow(clippy::too_many_arguments)]
 pub fn run_node(
@@ -50,7 +47,7 @@ pub fn run_node(
         ShowNodeMode::Both => NodeInfoOptions::both(),
     };
 
-    let node_infos = get_node_info(data, node_selector, predicates, options)?;
-    format_node_info_list(&node_infos, data, &mut writer)?;
+    let node_infos = get_node_info(data, node_selector, predicates, &options)?;
+    format_node_info_list(&node_infos, data, &mut writer, &options)?;
     Ok(())
 }