+++ rudof_cli/src/shex.rs
@@ -1,6 +1,5 @@
 use crate::data::get_data_rudof;
 use crate::data_format::DataFormat;
-use crate::node_selector::{parse_node_selector, parse_shape_selector, start};
 use crate::writer::get_writer;
 use crate::{ColorSupport, shapemap_format_convert, terminal_width};
 use crate::{ResultShExValidationFormat, ShapeMapFormat as CliShapeMapFormat};
@@ -9,7 +8,10 @@ use anyhow::Context;
 use anyhow::{Result, bail};
 use iri_s::IriS;
 use iri_s::mime_type::MimeType;
-use rudof_lib::{InputSpec, Rudof, RudofConfig, RudofError, ShExFormatter};
+use rudof_lib::{
+    InputSpec, Rudof, RudofConfig, RudofError, ShExFormatter, parse_node_selector,
+    parse_shape_selector, start,
+};
 use shex_ast::shapemap::{ResultShapeMap, ShapeSelector};
 use shex_ast::{Schema, ShExFormat};
 use srdf::{RDFFormat, ReaderMode};