+++ rudof_lib/src/lib.rs
@@ -3,18 +3,22 @@
 //!
 #![deny(rust_2018_idioms)]
 pub mod input_spec;
+pub mod node_formatter;
 pub mod node_info;
 pub mod rudof;
 pub mod rudof_config;
 pub mod rudof_error;
+pub mod selector;
 pub mod shapes_graph_source;
 
 pub use input_spec::*;
 pub use iri_s::*;
+pub use node_formatter::*;
 pub use oxrdf;
 pub use rudof::*;
 pub use rudof_config::*;
 pub use rudof_error::*;
+pub use selector::*;
 pub use shacl_ir;
 pub use shacl_validation;
 pub use shapes_graph_source::*;