+++ rudof_mcp/src/rudof_mcp_service/service.rs
@@ -15,7 +15,8 @@ pub struct RudofMcpService {
 impl RudofMcpService {
     pub fn new() -> Self {
         // TODO: Check and protect against possible initialization errors
-        let rudof = Rudof::new(&RudofConfig::new()).unwrap();
+        let rudof_config = RudofConfig::new().unwrap();
+        let rudof = Rudof::new(&rudof_config).unwrap();
         Self {
             rudof: Arc::new(Mutex::new(rudof)),
             tool_router: tools::tool_router_public(),